Disable hard-coded KVO variants from create-xcframeowkrs
Everything seems to be working well with this method… until building using a static target (did not try dynamic) where the build system complains two build phases produce the same files (the phase being the processing of the xcframework).
In an ideal world the build system would realize only one of the target is being used and would not try to process both, but we’re not in an ideal world, are we?